refactor: rewrite frontend with more readable format and components usage
